Ready set (#10)
* Add README.md (#4)

# Feature
This is a basic update to the readme file. More changes are definitely necessary but this is a good start. (Refer to #3) 
# Review
Check for spelling, comprehension, links, and completeness before merging.
## Satisfactory sections
- [x] - Project Name
- [x] - Description
- [x] - Table of contents
- [x] - Installation
- [x] - Contribution
- [x] - Milestones
- [x] - Credits
- [x] - License

* Create ISSUE_TEMPLATE.md (#8)

* Create ISSUE_TEMPLATE.md

Resolve #7.

* Update ISSUE_TEMPLATE.md

Included more descriptions to reporting bugs.

* Update ISSUE_TEMPLATE.md

Added checklist for reading CONTRIBUTING and CODE OF CONDUCT files.

Added section for describing user's environment